backend/requirements.txt
langchain==0.3.19
langchain-openai==0.3.7
langchain-core==0.3.39
langchain_community==0.3.18
docx2txt==0.8
pypdf==5.3.0
langchain_chroma==0.2.2
python-multipart==0.0.20
streamlit==1.42.0
sentence-transformers==2.7.0
pinecone-client==3.1.0
langchain-pinecone==0.2.2